Add more linters (#1788)

+++ internal/cmd/kafka/confluent_kafka_configs.go
@@ -264,7 +264,7 @@ func getOffsetWithFallback(cmd *cobra.Command) (ckafka.Offset, error) {
 func getPartitionsByIndex(partitions []ckafka.TopicPartition, partitionFilter partitionFilter) []ckafka.TopicPartition {
 	if partitionFilter.changed {
 		for _, partition := range partitions {
-			if partition.Partition == int32(partitionFilter.index) {
+			if partition.Partition == partitionFilter.index {
 				log.CliLogger.Debugf("Consuming from partition: %d", partitionFilter.index)
 				return []ckafka.TopicPartition{partition}
 			}
